Split up codegen code
