+++ package.json
@@ -110,7 +110,8 @@
     "collectCoverage": true,
     "collectCoverageFrom": [
       "<rootDir>/src/**/*.js",
-      "!<rootDir>/src/amphtml/amphtml.js"
+      "!<rootDir>/src/amphtml/amphtml.js",
+      "!<rootDir>/src/amphtml/codegen/**/*.js"
     ],
     "coverageThreshold": {
       "global": {
