+++ src/amphtml/codegen/constants.js
@@ -1,0 +1,10 @@
+module.exports.COMPONENT_OVERRIDES = {
+  AmpState: true,
+  Script: true,
+};
+
+module.exports.BLACKLIST = {
+  '!DOCTYPE': true,
+  $REFERENCE_POINT: true,
+  'O:P': true,
+};
