+++ src/amphtml/codegen/tagNameToComponentName.js
@@ -1,0 +1,8 @@
+module.exports = tagName => (
+  tagName
+    .toLowerCase()
+    .replace(
+      /(^.|-.)/g,
+      (m, p1) => p1.replace('-', '').toUpperCase(),
+    )
+);
